If the status of the task is 'Done', the icon for status will now be a check mark. Else, it will be icons.info
Task list:
for each task, it will display the title of the task in bold,
the due date
I removed the status since anything that has a check mark menas it done and anythign that has a different icon, has icons.info to show the status. So, user can quickly know whcih tasks are done.
I will add the option to go to more detail so that the user can see all details about a task